Commit producer (+update) spring boot 3.3.3 swagger ui Debug

+++ src/main/java/org/example/event/ProductEvent.java
@@ -2,6 +2,8 @@ package org.example.event;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @AllArgsConstructor
@@ -23,8 +25,8 @@ public class ProductEvent {
     private String businessAddress;
     private String address;
     private Integer rcbic;
-    private Double corrAss;
-    private Double ass;
+    private BigDecimal corrAss;
+    private BigDecimal ass;
     private String bankName;
     private Integer requisitesId;
 }
